LPD-30649 Extract VersionRow to its own component and convert it to Typescript
